% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips.R
\docType{data}
\name{fips}
\alias{fips}
\title{FIPS data}
\format{
\subsection{\code{fips}}{

A data frame with 3,204 rows and 5 columns:
\describe{
\item{statecode}{2-digit state FIPS code}
\item{countycode}{3-digit country FIPS codes}
\item{fipscode}{5-digit FIPS code}
\item{state}{state abbreviation}
\item{county}{county name}
}
}
}
\source{
\url{https://www.countyhealthrankings.org/sites/default/files/media/document}
}
\usage{
fips
}
\description{
A dataframe of FIPS codes for US counties.
}
\keyword{datasets}
